requirements.txt
# Core FastAPI and web framework dependencies
fastapi
uvicorn[standard]
python-dotenv

# Security and authentication
pydantic

# Database and vector storage
motor # MongoDB async driver
pymongo # MongoDB driver

# AI and ML dependencies
google-generativeai # Google AI Studio/Gemini
sentence-transformers # For embeddings
tiktoken # OpenAI tokenizer

# Notion API
notion-client

# HTTP client for external APIs
httpx
requests

# Testing dependencies (development)
pytest
pytest-asyncio
pytest-mock